+++ src/main.js
@@ -1,7 +1,8 @@
 import Vue from 'vue'
 import App from './App.vue'
 
-import './components';
+import KProgress from './components';
+Vue.component('k-progress', KProgress);
 
 Vue.config.productionTip = false