+++ CHANGELOG.md
@@ -1,13 +1,18 @@
 ## 📒 ChangeLog
+### v1.2
+`2019-12-28`
+- 💣 Attributes adjustment : `type` changed to `status`
+- 🎉 Added `border` to set whether the progress bar is circular
+- 💖 Added `lump` type
+  - 💜 Added `cut-width` to set `lump` width
+  - 🧡 Added `cut-color` to set `lump` color
 
 ### v1.1
 `2019-12-26`
-
-- 🎉 Progress bar color support `Function` setting
+- 🎉 Added progress bar color support `Function` setting
 
 ### v1.0
 `2019-12-18`
-
 - 🌟 Progress bar custom color
 - 🌟 Progress bar custom background color
 - 🌟 Custom height of progress bar